requirements.txt
asgiref==3.2.7
bson==0.5.8
certifi==2019.11.28
cloudinary==1.20.0
cycler==0.10.0
dask==2.14.0
dataclasses==0.6
decorator==4.4.2
Django==3.0.4
django-cors-headers==3.2.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
dnspython==1.16.0
kiwisolver==1.2.0
mongoengine==0.19.1
networkx==2.4
psycopg2==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
toolz==0.10.0
urllib3==1.25.8
gunicorn
mongomock
whitenoise
isbnlib